+++ src/Timer.php
@@ -12,13 +12,13 @@ namespace SebastianBergmann\Timer;
 final class Timer
 {
     /**
-     * @psalm-var list<int>
+     * @psalm-var list<float>
      */
     private $startTimes = [];
 
     public function start(): void
     {
-        $this->startTimes[] = \hrtime(true);
+        $this->startTimes[] = (float) \hrtime(true);
     }
 
     /**
@@ -32,6 +32,6 @@ final class Timer
             );
         }
 
-        return Duration::fromNanoseconds(\hrtime(true) - \array_pop($this->startTimes));
+        return Duration::fromNanoseconds((float) \hrtime(true) - \array_pop($this->startTimes));
     }
 }